+++ ref-exchange/tests/test_fuzzy_stable.rs
@@ -47,6 +47,7 @@ fn generate_fuzzy_seed() -> Vec<u64>{
 }
 
 #[test]
+#[ignore]
 fn test_fuzzy_stable() {
     let seeds = generate_fuzzy_seed();