+++ ref-exchange/release_notes.md
@@ -1,7 +1,7 @@
 # Release Notes
 
 ### Version 1.5.2
-1. Lower MIN_RESERVE to 1*10**16;
+1. Lower MIN_RESERVE to 1*10**15;
 
 ### Version 1.5.1
 1. Import rated stable pool;